ref: move FDPO to vocabulary package in rdf module
